requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0

# Box2D for CarRacing environment (use capital B and D for precompiled wheel)
Box2D>=2.3.10
pygame>=2.1.3

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Notebook Support
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
pyyaml>=6.0
